refactor out some repeated code
