+++ src/cron/helpers.ts
@@ -1,0 +1,20 @@
+import { db, getQueryResults } from '../db'
+
+export async function getActiveTeamsWithRehearsalOnDate(
+  date: Date,
+  extraWhereBool?: string
+) {
+  const day = date.getDay().toString()
+  const query = db
+    .collection('teams')
+    .where('rehearsal_day', '==', day)
+    .where('active', '==', true)
+
+  if (extraWhereBool) {
+    query.where(extraWhereBool, '==', true)
+  }
+
+  const teams = await getQueryResults(query)
+  if (teams.length === 0) return []
+  return teams
+}
